Implemented and tested refcounting on buffer objects
every bo now has a kref counter, counting works as follows:
- bo.refcnt is increased on any reference, that is not visible
  to userspace, especially vm mapping
- if the bo becomes user visible, it is wrapped by a gem, which
  increases bo.refcnt exactly by one
- the gem has its own, drm.refcnt = drm_handle_count + num_mmap
- drm_handle_count = 1 until user calls gem_close()
- when drm.refcnt == 0, bo.refcnt is decreased by one

The primary benefit is, that pscnv now handles bo's that are
mapped to more than one vspace with vid>0 correctly. This is
necessary, as DMA is handled by vspace 126 and needs to map
the bo.

A test (bo_refcnt) verifies that memory handling is correct.

/sys/kernel/debug/dri/0/memory now prints the actual vram usage.
There does not seem to be a memory leak