first revision of the "real" implementation
the data collected for the clients is now used for the scheduling
decision. Swaping operations get queued for the process that
owns the memory. Still the ioctl hack is needed to run the
queued functions.

pscnv_mem_alloc got another parameter (client responsible for
memory), that's why so many files had to be modified.