Much fancier build.sh and pscnv_load.sh scripts.
build.sh needs more work in terms of e.g. a command line parameter to
skip all the "make cleans" if one is sure that they are not needed.

pscnv_load.sh unloads any currently loaded GPU driver,
optionally performs reclocking via vbios patching and the blob,
and loads pscnv. For now, a fixed nvafakebios parameter string is used
when patching the bios. This may not work on other systems!
Ideally, the script should be improved to automatically decide
which bytes in the BIOS need to be replaced.

pscnv/Makefile now uses uname -r to set its SYSSRC again. IMHO,
any overrides to this should really be done from outside (e.g. the
build.sh script could be made to patch the Makefile as necessary)